+++ src/primitives/plane.c
@@ -6,13 +6,13 @@
 /*   By: bwerner <bwerner@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/06 16:00:59 by nmihaile          #+#    #+#             */
-/*   Updated: 2024/08/10 18:07:21 by bwerner          ###   ########.fr       */
+/*   Updated: 2024/08/10 20:14:58 by bwerner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/miniRT.h"
 
-static bool	is_near_zero(double value)
+static bool	is_near_zero(float value)
 {
 	return (value > -0.0001 && value < 0.0001);
 }
@@ -20,7 +20,7 @@ static bool	is_near_zero(double value)
 t_hitpoint	get_hitpoint_plane(t_ray ray, t_object *object)
 {
 	t_hitpoint	hitpoint;
-	double		dp;
+	float		dp;
 
 	if (is_near_zero(vec3_dot(vec3_sub(object->origin, ray.origin), object->normal)))
 		return (HP_INF);
@@ -33,7 +33,7 @@ t_hitpoint	get_hitpoint_plane(t_ray ray, t_object *object)
 		hitpoint.normal = object->normal;
 //	else if (backface culling is on)
 //		return (HP_INF);
-	double t = (object->dist - vec3_dot(ray.origin, object->normal)) / dp;
+	float t = (object->dist - vec3_dot(ray.origin, object->normal)) / dp;
 	if (t < 0)
 		return (HP_INF);
 	hitpoint.object = object;
