+++ src/utils/vec3_rotate.c
@@ -12,30 +12,30 @@
 
 #include "../../inc/miniRT.h"
 
-t_vec3	vec3_rotate_x(t_vec3 v, double rad)
+t_vec3	vec3_rotate_x(t_vec3 v, float rad)
 {
-	double	s;
-	double	c;
+	float	s;
+	float	c;
 
 	s = sin(rad);
 	c = cos(rad);
 	return ((t_vec3){v.x, v.y * c - v.z * s, v.y * s + v.z * c});
 }
 
-t_vec3	vec3_rotate_y(t_vec3 v, double rad)
+t_vec3	vec3_rotate_y(t_vec3 v, float rad)
 {
-	double	s;
-	double	c;
+	float	s;
+	float	c;
 
 	s = sin(rad);
 	c = cos(rad);
 	return ((t_vec3){v.x * c + v.z * s, v.y, -v.x * s + v.z * c});
 }
 
-t_vec3	vec3_rotate_z(t_vec3 v, double rad)
+t_vec3	vec3_rotate_z(t_vec3 v, float rad)
 {
-	double	s;
-	double	c;
+	float	s;
+	float	c;
 
 	s = sin(rad);
 	c = cos(rad);
@@ -52,7 +52,7 @@ t_vec3	vec3_rotate_z(t_vec3 v, double rad)
 // 	return (v);
 // }
 
-// t_vec3	vec3_rotate_x(t_vec3 p, double rad)
+// t_vec3	vec3_rotate_x(t_vec3 p, float rad)
 // {
 // 	t_vec3	v;
 // 	v = p;
@@ -72,7 +72,7 @@ t_vec3	vec3_rotate_z(t_vec3 v, double rad)
 // 	return (v);
 // }
 
-// t_vec3	vec3_rotate_y(t_vec3 p, double rad)
+// t_vec3	vec3_rotate_y(t_vec3 p, float rad)
 // {
 // 	t_vec3	v;
 // 	v = p;
@@ -92,7 +92,7 @@ t_vec3	vec3_rotate_z(t_vec3 v, double rad)
 // 	return (v);
 // }
 
-// t_vec3	vec3_rotate_z(t_vec3 p, double rad)
+// t_vec3	vec3_rotate_z(t_vec3 p, float rad)
 // {
 // 	t_vec3	v;
 // 	v = p;
